+++ cmd/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## [0.3.2]
-*
+* Bugfix: AzureResourceProperties.Other field saving to MongoDB as binary data
 
 ## [0.3.1]
 * Bugfix: Some resources showing in IP Address list when resource has no ip addresses