+++ cmd/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Bugfix: AzureResourceProperties.Other field saving to MongoDB as binary data
 * Add updateResources flag to mongodb.Update command
 * Add unsetField flag to mongodb.Clean command
+* Update deps
 
 ## [0.3.1]
 * Bugfix: Some resources showing in IP Address list when resource has no ip addresses