+++ cmd/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [0.3.2]
 * Bugfix: AzureResourceProperties.Other field saving to MongoDB as binary data
 * Add updateResources flag to mongodb.Update command
-* Add unsetField flag to mongodb.Clean command
+* Add unsetResourceField flag to mongodb.Clean command
 * Update deps
 
 ## [0.3.1]